+++ src/kernel/interrupt/interrupt.c
@@ -9,7 +9,7 @@ void set_idt_gate(unsigned char num,unsigned long base,unsigned short selector,u
 }
 void idt_install() {
 	idt_ptr.limit = (sizeof(idt_entry_t) * 256) - 1;
-	idt_ptr.base = &idt;
+	idt_ptr.base = (unsigned int)&idt;
 	memset(&idt,0,sizeof(idt_entry_t) * 256);
 	idt_load();
 }
